chore: rename file lightningcssMiminizerPlugin
