+++ packages/rspack/src/builtin-plugin/index.ts
@@ -9,7 +9,7 @@ export * from "./BundlerInfoRspackPlugin";
 export * from "./ChunkPrefetchPreloadPlugin";
 export * from "./CommonJsChunkFormatPlugin";
 export * from "./CopyRspackPlugin";
-export * from "./css-extract";
+export * from "./css-extract/index";
 export * from "./CssModulesPlugin";
 export * from "./DataUriPlugin";
 export * from "./DefinePlugin";
@@ -57,7 +57,7 @@ export * from "./SideEffectsFlagPlugin";
 export * from "./SizeLimitsPlugin";
 export * from "./SourceMapDevToolPlugin";
 export * from "./SplitChunksPlugin";
-export * from "./LightningCssMiminizerRspackPlugin";
+export * from "./LightningCssMinimizerRspackPlugin";
 export * from "./SwcJsMinimizerPlugin";
 export * from "./WarnCaseSensitiveModulesPlugin";
 export * from "./WebWorkerTemplatePlugin";
